Enhancement: Account center updates (#1740)
