+++ packages/demo/package.json
@@ -26,7 +26,7 @@
   "dependencies": {
     "@web3-onboard/cede-store": "^2.0.2",
     "@web3-onboard/coinbase": "^2.2.3",
-    "@web3-onboard/core": "^2.19.2",
+    "@web3-onboard/core": "^2.20.0-alpha.1",
     "@web3-onboard/dcent": "^2.2.7",
     "@web3-onboard/enkrypt": "^2.0.3",
     "@web3-onboard/fortmatic": "^2.0.18",
