+++ packages/demo/src/App.svelte
@@ -310,11 +310,10 @@
     // // example customizing account center
     accountCenter: {
       desktop: {
-        position: 'topRight',
+        position: 'topLeft',
         enabled: true,
-        minimal: false,
-      },
-      // hideTransactionProtectionBtn: true
+        minimal: true
+      }
     },
     // example customizing copy
     i18n: {
@@ -584,6 +583,10 @@
     flex-direction: column;
     width: 15rem;
   }
+  .position-buttons {
+    display: flex;
+    direction: row;
+  }
 </style>
 
 <main>
@@ -637,32 +640,6 @@
               )
             }}>Send Success Notification</button
           >
-          <button
-            on:click={() =>
-              onboard.state.actions.customNotification({
-                message:
-                  'This is a custom DApp success notification to use however you want',
-                autoDismiss: 0,
-                type: 'pending'
-              })}>Send Pending Notification</button
-          >
-          <button
-            on:click={() =>
-              onboard.state.actions.customNotification({
-                type: 'error',
-                message:
-                  'This is a custom DApp Error notification to use however you want',
-                autoDismiss: 0
-              })}>Send Error Notification</button
-          >
-          <button
-            on:click={() =>
-              onboard.state.actions.customNotification({
-                message:
-                  'This is a custom non-descript DApp notification to use however you want',
-                autoDismiss: 0
-              })}>Send DApp Notification</button
-          >
         </div>
         <div class="switch-chain-container">
           <button on:click={() => onboard.setChain({ chainId: '0x1' })}
@@ -678,6 +655,70 @@
             >Set Chain to Optimism</button
           >
         </div>
+      </div>
+      <div class="position-buttons">
+        <button
+          on:click={() =>
+            onboard.state.actions.customNotification({
+              message:
+                'This is a custom DApp success notification to use however you want',
+              autoDismiss: 0,
+              type: 'pending'
+            })}>Send Pending Notification</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.customNotification({
+              type: 'error',
+              message:
+                'This is a custom DApp Error notification to use however you want',
+              autoDismiss: 0
+            })}>Send Error Notification</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.customNotification({
+              message:
+                'This is a custom non-descript DApp notification to use however you want',
+              autoDismiss: 0
+            })}>Send DApp Notification</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.updateAccountCenter({
+              position: 'bottomLeft'
+            })}>AC Bottom Left</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.updateAccountCenter({
+              position: 'topRight'
+            })}>AC Top Right</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.updateAccountCenter({
+              position: 'bottomRight'
+            })}>AC Bottom Right</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.updateAccountCenter({
+              position: 'topLeft'
+            })}>AC Top Left</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.updateAccountCenter({
+              minimal: false
+            })}>Large Trigger</button
+        >
+        <button
+          on:click={() =>
+            onboard.state.actions.updateAccountCenter({
+              minimal: true
+            })}>Small Trigger</button
+        >
       </div>
     {/if}
   </div>
